refactor(brain): one primary type per file (review)
DetectedAgentCLI, AgentCLIRun, AgentCLIOutput, and BrainModel move into
their own files, named after the type, per the CLAUDE.md file-layout
rule — pure moves, no behavior change.
